Rename files using struct name
